Try to improve rendering time a bit
Only became noticable how slow it is with the filtering (500+ ms!)

Since we only see the *last* 10 elements in the crate at any given time,
put those in first and schedule the remainder to prepend after the next
render.

Really this is a bandaid.

The real culprit here is the cost of using a Polymer element with so
many new shadowroots created when a simple webcomponent would suffice.

The tragedy here is all my work to get the CSS effects to get along with
the shadowDOM ;~;